+++ ValheimServersideQoL/Processors/Processor.cs
@@ -442,6 +442,7 @@ abstract class Processor
 
     public static class Prefabs
     {
+        public static int Hammer { get; } = "Hammer".GetStableHashCode();
         public static int GraustenFloor4x4 { get; } = "Piece_grausten_floor_4x4".GetStableHashCode();
         public static int GraustenWall4x2 { get; } = "Piece_grausten_wall_4x2".GetStableHashCode();
         public static int PortalWood { get; } = "portal_wood".GetStableHashCode();